+++ src/pytest_workflow/__init__.py
@@ -15,6 +15,9 @@
 # along with pytest-workflow.  If not, see <https://www.gnu.org/licenses/
 
 import re
+import shutil
+from pathlib import Path
+from typing import List
 
 
 # This function was created to ensure the same conversion is used throughout
@@ -27,3 +30,8 @@ def replace_whitespace(string: str, replace_with: str = '_') -> str:
     :return: The string with whitespace converted.
     """
     return re.sub(r'\s+', replace_with, string)
+
+
+def rm_dirs(directories: List[Path]):
+    for directory in directories:
+        shutil.rmtree(str(directory))