+++ tests/test_temp_directory.py
@@ -25,7 +25,7 @@ from .test_success_messages import SIMPLE_ECHO
 def test_directory_kept(testdir):
     testdir.makefile(".yml", test=SIMPLE_ECHO)
     result = testdir.runpytest("-v", "--keep-workflow-wd")
-    working_dir = re.search(r"with command 'echo moo' in '([\w\/_-]*)'",
+    working_dir = re.search(r"with command 'echo moo' in '([\w/_-]*)'",
                             result.stdout.str()).group(1)
     assert Path(working_dir).exists()
     assert Path(working_dir / Path("log.out")).exists()
@@ -35,7 +35,7 @@ def test_directory_kept(testdir):
 def test_directory_not_kept(testdir):
     testdir.makefile(".yml", test=SIMPLE_ECHO)
     result = testdir.runpytest("-v")
-    working_dir = re.search(r"with command 'echo moo' in '([\w\/_-]*)'",
+    working_dir = re.search(r"with command 'echo moo' in '([\w/_-]*)'",
                             result.stdout.str()).group(1)
     assert not Path(working_dir).exists()
     assert ("Removing temporary directories and logs. Use '--kwd' or "
@@ -83,3 +83,49 @@ def test_basetemp_will_be_created(testdir):
                                str(tempdir))
     assert tempdir.exists()
     assert result.ret == 0
+
+
+def test_basetemp_can_not_be_in_rootdir(testdir):
+    testdir.makefile(".yml", test=SIMPLE_ECHO)
+    testdir_path = Path(str(testdir.tmpdir))
+    tempdir = testdir_path / Path("tmp")
+    result = testdir.runpytest("-v", "--basetemp", str(tempdir))
+    message = "'{tempdir}' is a subdirectory of '{rootdir}'".format(
+        tempdir=str(tempdir),
+        rootdir=str(testdir_path))
+    assert message in result.stderr.str()
+
+
+SUCCESS_TEST = """\
+- name: success
+  command: bash -c 'exit 0'
+"""
+
+FAIL_TEST = """\
+- name: fail
+  command: bash -c 'exit 1'
+"""
+
+
+def test_directory_kept_on_fail(testdir):
+    testdir.makefile(".yml", test=FAIL_TEST)
+    result = testdir.runpytest("-v", "--keep-workflow-wd-on-fail")
+    working_dir = re.search(
+        r"with command 'bash -c 'exit 1'' in '([\w/_-]*)'",
+        result.stdout.str()).group(1)
+    assert Path(working_dir).exists()
+    assert Path(working_dir / Path("log.out")).exists()
+    assert Path(working_dir / Path("log.err")).exists()
+    assert ("One or more tests failed. Keeping temporary directories and "
+            "logs." in result.stdout.str())
+
+
+def test_directory_not_kept_on_succes(testdir):
+    testdir.makefile(".yml", test=SUCCESS_TEST)
+    result = testdir.runpytest("-v", "--kwdof")
+    working_dir = re.search(
+        r"with command 'bash -c 'exit 0'' in '([\w/_-]*)'",
+        result.stdout.str()).group(1)
+    assert not Path(working_dir).exists()
+    assert ("All tests succeeded. Removing temporary directories and logs." in
+            result.stdout.str())