+++ HISTORY.rst
@@ -7,6 +7,15 @@ Changelog
 .. NOTE: This document is user facing. Please word the changes in such a way
 .. that users understand how the changes affect the new version.
 
+version 1.2.0
+---------------------------
++ Giving a ``--basetemp`` directory that is within pytest's current working
+  directory will now raise an exception to prevent infinite recursive directory
+  copying.
++ The cleanup message is only displayed when pytest-workflow is used.
++ Added a ``--keep-workflow-wd-on-fail`` or ``--kwdof`` flag. Setting this flag
+  will make sure temporary directories are only deleted when all tests succeed.
+
 version 1.1.2
 ---------------------------
 + Fixed a bug where the program would hang indefinitely after a user input